Moves, rename and add version to api controller. moves api/email_controller to api/v1/email_requests
